estimate gas in transacting
